move doAutoRangeAndConstraints, drawData and finalDraw to subroutines
... (i.e. out of plot_api.js),
    and move Plots.addLinks out of drawData into its own
    step in Plotly.plot.
